backend/foodgram/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==23.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.4
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.3
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
psycopg2-binary==2.9.3
